fix: board controller, service, module
controller url 변경
service 예외처리 추가
module imports workspace 추가

+++ src/boards/boards.service.ts
@@ -15,7 +15,9 @@ export class BoardsService {
 
   // 보드 조회
   async GetBoards(workspaceId: number) {
+    const workspace = await this.workspaceRepository.findOneBy({ id: workspaceId });
     const findBoards = await this.boardRepository.find({ relations: ['workspace'] });
+    if (!workspace) throw new NotFoundException('해당 워크스페이스는 존재하지 않습니다.');
 
     return findBoards.filter((board) => {
       return board.workspace.id == workspaceId;
@@ -24,12 +26,16 @@ export class BoardsService {
 
   //보드 상세 조회
   async GetBoardById(workspaceId: number, id: number) {
+    const workspace = await this.workspaceRepository.findOneBy({ id: workspaceId });
+    if (!workspace) throw new NotFoundException('해당 워크스페이스는 존재하지 않습니다.');
+
     return await this.boardRepository.findOneBy({ id });
   }
 
   //보드 생성
   async CreateBoard(workspaceId: number, name: string, description: string) {
     const workspace = await this.workspaceRepository.findOneBy({ id: workspaceId });
+    if (!workspace) throw new NotFoundException('해당 워크스페이스는 존재하지 않습니다.');
     if (!name || !description) throw new NotFoundException('데이터 형식이 올바르지 않습니다.');
 
     await this.boardRepository.insert({ name, description, workspace });
@@ -37,12 +43,20 @@ export class BoardsService {
 
   //보드 수정
   async UpdateBoard(workspaceId: number, id: number, name: string, description: string) {
+    const workspace = await this.workspaceRepository.findOneBy({ id: workspaceId });
+    const board = await this.boardRepository.findOneBy({ id });
+    if (!workspace) throw new NotFoundException('해당 워크스페이스는 존재하지 않습니다.');
+    if (!board) throw new NotFoundException('해당 보드는 존재하지 않습니다.');
     if (!name || !description) throw new NotFoundException('데이터 형식이 올바르지 않습니다.');
     await this.boardRepository.update({ id }, { name, description });
   }
 
   //보드 삭제
   async DeleteBoard(workspaceId: number, id: number) {
+    const workspace = await this.workspaceRepository.findOneBy({ id: workspaceId });
+    const board = await this.boardRepository.findOneBy({ id });
+    if (!workspace) throw new NotFoundException('해당 워크스페이스는 존재하지 않습니다.');
+    if (!board) throw new NotFoundException('해당 보드는 존재하지 않습니다.');
     await this.boardRepository.delete(id);
   }
 }
